+++ NEWS.md
@@ -6,6 +6,9 @@
 
 * Parameters were added for the `tab_pfn` model: `num_estimators()`, `softmax_temperature()`, `balance_probabilities()`, `average_before_softmax()`, and `training_set_limit()`.
 
+* `parameters()` and the `grid_*()` functions give more information in the error message when non-parameter objects are passed in (#437).
+
+
 # dials 1.4.2
 
 * `prop_terms()` is a new parameter object used for recipes that do supervised feature selection (#395).